+++ src/test/testData/state/outsideFactory/essential/propertyWithNullableType.php
@@ -1,0 +1,22 @@
+<?php
+
+class Account {
+    public function __construct(
+        public readonly string <caret>$id,
+        public readonly ?int $age,
+    ) {}
+}
+
+class AccountFactory extends EKvedaras\ClassFactory\ClassFactory {
+    protected string $class = Account::class;
+
+    protected function definition(): array
+    {
+        return [
+            'id' => 'abc',
+            'age' => 1,
+        ];
+    }
+}
+
+AccountFactory::new()->state(['age' => null]);